% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balancedstratification2.R
\name{balancedstratification2}
\alias{balancedstratification2}
\title{Title}
\usage{
balancedstratification2(X, strata, pik)
}
\arguments{
\item{X}{}

\item{strata}{}

\item{pik}{}

\item{comment}{}

\item{method}{}
}
\value{

}
\description{
Title
}
\examples{

rm(list = ls())
N <- 10000
n <- c(25,50,100,250)
p <- 3
X_tmp <- matrix(rgamma(N*p,4,25),ncol = p)

l <- list()
sizeStrata <- c(1,1.5)
step <- 1

i = 4
j = 1
strata <- as.matrix(rep(1:n[i],each = N/n[i]))
pik <- inclusionprobastrata(strata,rep(sizeStrata[j],n[i]))
X <- cbind(pik,X_tmp)
comment = TRUE

balancedstratification2(X,strata,pik,comment=TRUE))[3]

}
