% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balcat.R
\name{balcat}
\alias{balcat}
\title{Title}
\usage{
balcat(X, Xcat, pik)
}
\arguments{
\item{pik}{}
}
\value{

}
\description{
Title
}
\examples{


rm(list = ls())
N <- 100
pik <- rep(10/N,N)

Xcat <-data.frame(cat1 = rep(1:50,each = 2))
Xcat_tmp <- disjMatrix(as.matrix(Xcat))
X <- cbind(pik,matrix(rnorm(N),ncol = 1))
A <- cbind(X,Xcat_tmp)/pik

s <- balcat(X,Xcat,pik)
t(A)\%*\%s
t(A)\%*\%pik


rm(list = ls())
N <- 10000
pik <- rep(0.25,N)

Xcat <- as.matrix(rep(1:1250,each = N/1250))
Xcat_tmp <- disjMatrix(Xcat)
X <- cbind(pik,matrix(rnorm(N),ncol = 1))
A <- cbind(X,Xcat_tmp)/pik

system.time(s <- balcat(X,Xcat,pik))



rm(list = ls())
N <- 1000
pik <- rep(0.25,N)

Xcat <- as.matrix(rep(1:250,each = N/250))
Xcat_tmp <- disjMatrix(Xcat)
X <- cbind(pik,matrix(rnorm(N),ncol = 1))
A <- cbind(X,Xcat_tmp)/pik


system.time(s <- balcat(X,Xcat,pik))
system.time(s <- BalancedSampling::flightphase(pik,cbind(X,Xcat_tmp)))
system.time(s <- balancedstratification(X,Xcat,pik,comment=FALSE,method=2))

microbenchmark( balcat(X,Xcat,pik),
BalancedSampling::flightphase(pik,cbind(X,Xcat_tmp)),
balancedstratification(X,Xcat,pik,comment=FALSE,method=2),times = 10)



rm(list = ls())
N <- 120
pik <- rep(0.4,N)
pikInit <- pik
Xcat <-data.frame(cat1 = rep(1:40,each = 3))
Xcat_tmp <- disjMatrix(as.matrix(Xcat))
X <- cbind(pik,matrix(rnorm(N),ncol = 1))
X <- as.matrix(pik)
A <- cbind(X,Xcat_tmp)/pik

s <- balcat(X,Xcat,pik)

t(A)\%*\%s
t(A)\%*\%pik



rm(list = ls())
N <- 10000
n <- 1000
x1 <- rgamma(N,4,25)
x2 <- rgamma(N,4,25)

Xcat <- as.matrix(rep(1:n,each = N/n))

pik <- inclusionprobastrata(Xcat,rep(1,n))
#pik <- rep(inclusionprobabilities(runif(400),1),25)

X <- as.matrix(cbind(matrix(c(x1,x2),ncol = 2)))

system.time(s <- balcat(X,Xcat,pik))

}
