% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balstrat.R
\name{balstrat}
\alias{balstrat}
\title{Balanced Stratification}
\usage{
balstrat(X, strata, pik)
}
\arguments{
\item{X}{matrix of auxiliary variables.}

\item{strata}{vector of integer that specifies the stratification.}

\item{pik}{vector of inclusion probabilities.}
}
\value{
A vector with elements equal to 0 or 1. The value 1 indicates that the unit is selected while the value 0 is for rejected units.
}
\description{
Select a strafitifed balanced sample. The function is similar to \code{\link[sampling:balancedstratification]{balancedstratification}} unless it uses internal functions that are more stable.
}
\examples{

N <- 100
n <- 10
p <- 4
X <- matrix(rgamma(N*p,4,25),ncol = p)
strata <- as.matrix(rep(1:n,each = N/n))
pik <- rep(n/N,N)

s <- balstrat(X,strata,pik)

t(X/pik)\%*\%s
t(X/pik)\%*\%pik

Xcat <- disj(strata)

t(Xcat)\%*\%s
t(Xcat)\%*\%pik
}
