% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{c_bound2}
\alias{c_bound2}
\title{C bound}
\usage{
c_bound2(pik)
}
\arguments{
\item{pik}{vector of the inclusion probabilities.}
}
\value{
An integer value, the number of units that we need to respect the constraints.
}
\description{
This function is returning the number of unit that we need such that some conditions are fulfilled. See Details
}
\details{
The function is computing the number of unit \eqn{K} that we need to add such that the following conditions are fulfilled :

\itemize{
\item \eqn{\sum_{k = 1}^K \pi_k \geq 1}
\item \eqn{\sum_{k = 1}^K 1 - \pi_k \geq 1}
\item Let \eqn{c} be the constant such that \eqn{\sum_{k = 2}^K min(c\pi_k,1) = n }, we must have that \eqn{ \pi_1 \geq 1- 1/c}
}
}
\seealso{
\code{\link{osod}}
}
\author{
Raphael Jauslin \email{raphael.jauslin@unine.ch}
}
