% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{choose}
\alias{choose}
\title{Binomial Coefficient}
\usage{
choose(n, k)
}
\arguments{
\item{n}{An integer. (greater than k)}

\item{k}{An integer.}
}
\value{
An integer "n choose k".
}
\description{
It calculates the number of sets with size \eqn{k} that can be chosen from a set of size \eqn{n}.
}
\examples{
choose(10,5)
}
\seealso{
\code{\link[base:choose]{base::choose}}
}
\author{
Raphael Jauslin \email{raphael.jauslin@unine.ch}
}
