% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{disj}
\alias{disj}
\title{Disjunctive}
\usage{
disj(strata)
}
\arguments{
\item{strata}{A vector of integers that represents the categories.}
}
\value{
A matrix of indicators.
}
\description{
This function transforms a categorical vector into a matrix of indicators.
}
\examples{
strata <- rep(c(1,2,3),each = 4)
disj(strata)

}
\author{
Raphaël Jauslin \email{raphael.jauslin@unine.ch}
}
