% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{disj_to_cat}
\alias{disj_to_cat}
\title{findBarma}
\usage{
disj_to_cat(Xcat)
}
\arguments{
\item{Xcat}{a Matrix}

\item{X}{a Matrix}
}
\value{
a matrix
}
\description{
findB
}
\author{
Raphaël Jauslin \email{raphael.jauslin@unine.ch}
}
