% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ffphase_cpp}
\alias{ffphase_cpp}
\title{title}
\usage{
ffphase_cpp(X, pik, EPS = 1e-07)
}
\arguments{
\item{X}{matrix of auxiliary variables.}

\item{pik}{vector of inclusion probabilities}

\item{EPS}{tolerance}
}
\value{
a sample
}
\description{
title
}
\author{
Raphaël Jauslin \email{raphael.jauslin@unine.ch}
}
