% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ffphase_graf}
\alias{ffphase_graf}
\title{Fast Flight phase}
\usage{
ffphase_graf(prob, Xbal, order = TRUE, redux = FALSE)
}
\arguments{
\item{prob}{vector of inclusion probabilities of size N.}

\item{Xbal}{Matrix of auxiliary variables of dimension N x p}

\item{order}{if reordering at first step, Default TRUE.}

\item{redux}{if the matrix should be reduced. Default FALSE.}
}
\value{
a sample with at most p value not update to 0 or 1.
}
\description{
Modified version of \code{\link[BalancedSampling:flightphase]{flightphase}}
}
\details{
details
}
\author{
Raphaël Jauslin \email{raphael.jauslin@unine.ch}
}
