% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landing.R
\name{landingRM}
\alias{landingRM}
\title{Landing by suppression of variables}
\usage{
landingRM(X, pikstar, pik)
}
\arguments{
\item{X}{matrix of auxiliary variables on which the sample must be balanced.}

\item{pikstar}{vector of updated inclusion probabilities by the flight phase.}

\item{pik}{vector of inclusion probabilities.}

\item{Xcat}{matrix of categorical variable (not in its disjunctive form).}
}
\description{
This function does the landing phase of the cube method using suppression of variables.
 It allows you to put some categorical variable using the \code{Xcat} variable.
}
\examples{
\dontrun{
rm(list = ls())
N <- 100
n <- 10
p <- 4

pik <- sampling::inclusionprobabilities(runif(N),n)

X <- cbind(pik,matrix(rnorm(N*p),ncol= p))
pikstar <- ffphase(X,pik) 
s <- landingRM(X,pikstar,pik)
sum(s)
t(X/pik)\%*\%pik
t(X/pik)\%*\%pikstar
t(X/pik)\%*\%s
}
}
\references{
Chauvet, G. and Tille, Y. (2006). A fast algorithm of balanced sampling. Computational Statistics, 21/1:53-62
}
\author{
Raphaël Jauslin \email{raphael.jauslin@unine.ch}
}
