% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{maxent}
\alias{maxent}
\title{Maximum entropy sampling}
\usage{
maxent(pikr)
}
\arguments{
\item{pikr}{A vector of inclusion probabilities.}
}
\value{
A vector with elements equal to 0 or 1. The value 1 indicates that the unit is selected while the value 0 is for rejected units.
}
\description{
Maximum entropy sampling with fixed sample size. It can handle unequal inclusion probabilities.
}
\details{
The sampling design maximizes the entropy design:
\deqn{I(p) = - \sum s p(s) log[p(s)].}

This function is a C++ implementation of \code{\link[sampling:UPmaxentropy]{UPmaxentropy}}.
More details could be find in Tille (2006).
}
\references{
Tille, Y. (2006), Sampling Algorithms, springer
}
\author{
Raphaël Jauslin \email{raphael.jauslin@unine.ch}
}
