% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pikfromq}
\alias{pikfromq}
\title{pik from q}
\usage{
pikfromq(q)
}
\arguments{
\item{q}{A matrix that is computed from the function \code{\link{qfromw}}.}
}
\value{
A vector of inclusion probability computed from the matrix \code{q}.
}
\description{
This function finds the \code{pik} from an initial \code{q}.
}
\details{
More details could be find in Tille (2006).
}
\references{
Tille, Y. (2006), Sampling Algorithms, springer
}
\author{
Raphaël Jauslin \email{raphael.jauslin@unine.ch}
}
