% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{piktfrompik}
\alias{piktfrompik}
\title{pikt from pik}
\usage{
piktfrompik(pik, max_iter = 500L, tol = 1e-08)
}
\arguments{
\item{pik}{A vector of inclusion probabilities. The vector must contains only value that are not integer.}

\item{max_iter}{An integer that specify the maximum iteration in the Newton-Raphson algorithm. Default \code{500}.}

\item{tol}{A scalar that specify the tolerance convergence for the Newton-Raphson algorithm. Default \code{1e-8}.}
}
\value{
An updated vector of inclusion probability.
}
\description{
This function finds the \code{pikt} from an initial \code{pik}.
}
\details{
The management of probabilities equal to 0 or 1 is done in the cps function.

\code{pikt} is the vector of inclusion probabilities of a Poisson sampling with the right parameter. The vector is found by Newtwon-Raphson algorithm.

More details could be find in Tille (2006).
}
\references{
Tille, Y. (2006), Sampling Algorithms, springer
}
\author{
Raphaël Jauslin \email{raphael.jauslin@unine.ch}
}
