% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{qfromw}
\alias{qfromw}
\title{q from w}
\usage{
qfromw(w, n)
}
\arguments{
\item{w}{A vector of weights.}

\item{n}{An integer that is equal to the sum of the inclusion probabilities.}
}
\value{
A matrix of size \code{N} x \code{n}, where \code{N} is equal to the length of the vector \code{w}.
}
\description{
This function finds the matrix \code{q} form a particular \code{w}.
}
\details{
\code{w} is generally computed by the formula \code{pik/(1-pik)}, where \code{n} is equal to the sum of the vector \code{pik}.
More details could be find in Tille (2006).
}
\references{
Tille, Y. (2006), Sampling Algorithms, springer
}
\author{
Raphaël Jauslin \email{raphael.jauslin@unine.ch}
}
