% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{samplen}
\alias{samplen}
\title{Samples of fixed size}
\usage{
samplen(N, n)
}
\arguments{
\item{N}{An integer, the size of the population}

\item{n}{An integer, the sample size.}
}
\value{
A matrix of indicator variables.
}
\description{
Gives a matrix that contains in each column a possible sample of size \eqn{n}.
}
\details{
It uses a fast implementation. See References.
}
\examples{
samplen(4,2)

}
\references{
\url{http://www.rosettacode.org/wiki/Rosetta_Code}
}
\seealso{
\code{\link{landingLP}}
}
\author{
Raphaël Jauslin \email{raphael.jauslin@unine.ch}
}
