% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sfromq}
\alias{sfromq}
\title{s from q}
\usage{
sfromq(q)
}
\arguments{
\item{q}{A matrix that is computed from the function \code{\link{qfromw}}.}
}
\value{
A vector with elements equal to 0 or 1. The value 1 indicates that the unit is selected while the value 0 is for rejected units.
}
\description{
This function finds sample \code{s} form the matrix \code{q}.
}
\details{
More details could be find in Tille (2006).
}
\references{
Tille, Y. (2006), Sampling Algorithms, springer
}
\author{
Raphaël Jauslin \email{raphael.jauslin@unine.ch}
}
