% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SwissCommune}
\alias{SwissCommune}
\title{Shapefile of municipalities of Switzerland}
\format{
A data frame with 1559 rows and 5 variables:
\describe{
  \item{\code{Time (s)}}{double COLUMN_DESCRIPTION}
  \item{\code{Acceleration - x (m/s²)}}{double COLUMN_DESCRIPTION}
  \item{\code{Acceleration - y (m/s²)}}{double COLUMN_DESCRIPTION}
  \item{\code{Acceleration - z (m/s²)}}{double COLUMN_DESCRIPTION}
  \item{\code{Acceleration - resultant (m/s²)}}{double COLUMN_DESCRIPTION} 
}
}
\source{
\url{https://www.swisstopo.admin.ch/}
}
\usage{
SwissCommune
}
\description{
balbal
}
\keyword{datasets}
