% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varStrat.R
\name{varApp}
\alias{varApp}
\title{Title}
\usage{
varApp(X, strata, pik, y)
}
\arguments{
\item{X}{}

\item{strata}{}

\item{pik}{}

\item{s}{}
}
\value{

}
\description{
Title
}
\examples{


rm(list = ls())
N <- 1000
n <- 400
x1 <- rgamma(N,4,25)
x2 <- rgamma(N,4,25)

strata <- as.matrix(rep(1:40,each = 25)) # 25 strata
Xcat <- disjMatrix(strata)

#-------- CASE 1 pik equal and integer in each strata 

pik <- inclusionprobastrata(strata,rep(1,n))
X <- as.matrix(matrix(c(x1,x2),ncol = 2))
 
s <- stratifiedcube(cbind(pik,X),strata,pik)
 
 
 y <- 500 + 5*x1 + 5*x2 + rnorm(1000,0,270)
 
 y_ht <- sum(y[which(s==1)]/pik[which(s == 1)])
 
 sum(y)
 sum(y_ht)
 
 varApp(X,strata,pik,s,y)
 
}
