% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varStrat.R
\name{varEst}
\alias{varEst}
\title{Title}
\usage{
varEst(X, strata, pik, s, y)
}
\arguments{
\item{X}{}

\item{strata}{}

\item{pik}{}

\item{s}{}
}
\value{

}
\description{
Title
}
\examples{


rm(list = ls())
N <- 1000
n <- 400
x1 <- rgamma(N,4,25)
x2 <- rgamma(N,4,25)

strata <- as.matrix(rep(1:40,each = 25)) # 25 strata
Xcat <- disjMatrix(strata)

#-------- CASE 1 pik equal and integer in each strata 

pik <- rep(n/N,N)
X <- as.matrix(matrix(c(x1,x2),ncol = 2))
 
s <- stratifiedcube(cbind(pik,X),strata,pik)
s <- fbs(X,strata,pik)
 
 y <- 20*strata + rnorm(1000,120)
 y2 <- 500 + 5*x1 + 5*x2 + rnorm(1000,0,270)
 
 y_ht <- sum(y[which(s==1)]/pik[which(s == 1)])
 
 (sum(y_ht) - sum(y))^2
 varEst(X,strata,pik,s,y)
 varApp(X,strata,pik,s,y)
 
}
