% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wosi.R
\name{wosi}
\alias{wosi}
\title{Windows of size Integer}
\usage{
wosi(pik, H)
}
\arguments{
\item{pik}{A vector of inclusion probabilities.}

\item{H}{An integer, the considered windows size.}
}
\value{
A vector with elements equal to 0 or 1. The value 1 indicates that the unit is selected while the value 0 is for rejected units.
}
\description{
Windows of size Integer
}
\examples{
pik<-c(.4,.7,.1,.2,.4,.2,.2,.2,.1,.6,.2,.1,.4,.5,.5,.6,.1,.2)
H <- 2
s <- wosi(pik,2)

}
