% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wosi.R
\name{wosi2}
\alias{wosi2}
\title{wosi}
\usage{
wosi2(pik, H)
}
\arguments{
\item{pik}{}

\item{H}{}
}
\value{

}
\description{
wosi
}
\examples{
rm(list = ls())

library(sampling)
pik<-c(.4,.7,.1,.2,.4,.2,.2,.2,.1,.6,.2,.1,.4,.5,.5,.6,.1,.2)
pik <- inclusionprobabilities(runif(100),10)
H <- 2
s <- wosi2(pik,H)
s
sum(s)
}
