% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{wosicpp}
\alias{wosicpp}
\title{wosi}
\usage{
wosicpp(pikr, H)
}
\arguments{
\item{pikr}{A vector of inclusion probabilities.}
}
\value{
A vector with elements equal to 0 or 1. The value 1 indicates that the unit is selected while the value 0 is for rejected units.
}
\description{
sakdhfa
}
\details{

}
\examples{

N <- 1000
n <- 100
pik <- inclprob(runif(N),n)
s <- osod(pik)

}
\author{
Raphael Jauslin \email{raphael.jauslin@unine.ch}
}
