% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{disjMatrix}
\alias{disjMatrix}
\title{Disjunctive for matrix}
\usage{
disjMatrix(strata)
}
\arguments{
\item{strata}{A matrix of integers that contains categorical vector in each column.}
}
\value{
A matrix of indicators.
}
\description{
This function transforms a categorical matrix into a matrix of indicators variables.
}
\examples{
Xcat <-  matrix(c(sample(x = 1:6, size = 100, replace = TRUE),
            sample(x = 1:6, size = 100, replace = TRUE),
            sample(x = 1:6, size = 100, replace = TRUE)),ncol = 3)
disjMatrix(Xcat)

}
\author{
Raphaël Jauslin \email{raphael.jauslin@unine.ch}
}
