% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fbs.R
\name{fbs}
\alias{fbs}
\title{Title}
\usage{
fbs(X, Xcat, pik)
}
\arguments{
\item{pik}{}
}
\value{

}
\description{
Title
}
\examples{
rm(list = ls())
N <- 10000
n <- 1000
x1 <- rgamma(N,4,25)
x2 <- rgamma(N,4,25)

Xcat <- as.matrix(rep(1:n,each = N/n))

pik <- inclusionprobastrata(Xcat,rep(1,n))
X <- as.matrix(cbind(matrix(c(x1,x2),ncol = 2)))

A <- cbind(X,disjunctive(as.matrix(Xcat)))

system.time(s <- fbs(X,Xcat,pik))

}
