% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fbs.R
\name{fbs}
\alias{fbs}
\title{Fast balanced Sampling}
\usage{
fbs(X, Xcat, pik)
}
\arguments{
\item{X}{matrix of auxiliary variables.}

\item{Xcat}{matrix of categorical variables.}

\item{pik}{vector of inclusion probabilities.}
}
\value{
A vector with elements equal to 0 or 1. The value 1 indicates that the unit is selected while the value 0 is for rejected units.
}
\description{
This function implements the method proposed by Hasler and Tille (2019). It should be used for selecting a sample for highly stratified population.
}
\details{
If the number of element selected in each strata is not equal to a integer, the function can be very time-consuming.
}
\examples{

N <- 1000
n <- 100
x1 <- rgamma(N,4,25)
x2 <- rgamma(N,4,25)

Xcat <- as.matrix(rep(1:n,each = N/n))

pik <- sampling::inclusionprobastrata(Xcat,rep(1,n))
X <- as.matrix(cbind(matrix(c(x1,x2),ncol = 2)))
A <- cbind(X,sampling::disjunctive(as.matrix(Xcat)))

s <- fbs(X,Xcat,pik)

}
\references{
Hasler, C. and Tille Y. (2014). Fast balanced sampling for highly stratified population.
\emph{Computational Statistics and Data Analysis}, 74, 81-94
}
