% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffphase.R
\name{ffphase}
\alias{ffphase}
\title{Fast flight phase}
\usage{
ffphase(X, pik)
}
\arguments{
\item{X}{a matrix of size (N x p) of auxiliary variables on which the sample must be balanced.}

\item{pik}{a vector of size N of inclusion probabilities.}
}
\value{
the updated vector of \code{pik} that contains 0 and 1 for unit that are rejected or selected.
}
\description{
This function is computing the flight phase of the cube method.
}
\details{
This function implements the method proposed by (Chauvet and Tille 2006). It progressively transform the vector of inclusion probabilities \code{pik} into a sample while respecting the balancing equations.
The algorithm stops when the null space of the matrix \eqn{B} is empty. For more information see (Chauvet and Tille 2006).

The function uses the function \code{\link[MASS:Null]{Null}} to find the null space of the matrix \eqn{B}.
}
\examples{

N <- 100
n <- 10
p <- 4

pik <- rep(n/N,N)
X <- cbind(pik,matrix(rnorm(N*p),ncol= p))

pikstar <- ffphase(X,pik) 

t(X/pik)\%*\%pikstar
t(X/pik)\%*\%pik
pikstar

}
\references{
Chauvet, G. and Tille, Y. (2006). A fast algorithm of balanced sampling. Computational Statistics, 21/1:53-62
}
\seealso{
\code{\link[sampling:samplecube]{fastflightphase}}, \code{\link[BalancedSampling:flightphase]{flightphase}}.
}
\author{
Raphaël Jauslin \email{raphael.jauslin@unine.ch}
}
