% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findB.R
\name{findB}
\alias{findB}
\title{find the sub-matrix B in the flight phase}
\usage{
findB(X, strata)
}
\arguments{
\item{X}{matrix of auxiliary variables.}

\item{Xcat}{matrix of categorical variables.}
}
\value{
matrix B
}
\description{

}
\details{
The function find the smallest matrix B such that is contains only one additional row.
 It consecutively add the right number of row depending of the number of categroies that is added.
}
\examples{
N <- 1000
strata <-  sample(x = 1:6, size = N, replace = TRUE)

p <- 3
X <- matrix(rnorm(N*p),ncol = 3)
findB(X,strata)


N <- 100
Xcat <-data.frame(cat1 = rep(1:50,each = 2))
Xcat <- as.matrix(Xcat[-2,])
Xcat <- as.matrix(Xcat[-4,])
X <- matrix(rnorm(N),ncol = 1)

findB(X,Xcat)



}
