% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratifiedcube.R
\name{stratifiedcube}
\alias{stratifiedcube}
\title{Title}
\usage{
stratifiedcube(X, strata, pik)
}
\arguments{
\item{pik}{}
}
\value{

}
\description{
Title
}
\examples{

##----------------------------------------------------------------
##                              Data                             -
##----------------------------------------------------------------

rm(list = ls())
N <- 10000
n <- 1000
x1 <- rgamma(N,4,25)
x2 <- rgamma(N,4,25)
strata <- as.matrix(rep(1:n,each = N/n))
Xcat <- disjMatrix(strata)



##---------------------------------------------------------------
##                        different cases                       -
##---------------------------------------------------------------


 #-------- CASE 0 pik equal and only pik as variable
 
 pik <- inclusionprobastrata(strata,rep(1,n))
 X <- as.matrix(pik)
 system.time(s <- stratifiedcube(X,strata,pik))
 sum(s)
 t(X/pik)\%*\%s
 t(X/pik)\%*\%pik
 
 t(Xcat)\%*\%s
 t(Xcat)\%*\%pik
 
#-------- CASE 0.1 pik equal and only pik as variable
 
 pik <- rep(inclusionprobabilities(runif(N/n),1),n)
 X <- as.matrix(pik)
 system.time(s <- stratifiedcube(X,strata,pik))
 sum(s)
 t(X/pik)\%*\%s
 t(X/pik)\%*\%pik
 
 t(Xcat)\%*\%s
 t(Xcat)\%*\%pik
  

#-------- CASE 1 pik equal and integer in each strata 

 pik <- inclusionprobastrata(strata,rep(1,n))
 X <- as.matrix(cbind(pik,matrix(c(x1,x2),ncol = 2)))
 system.time(s <- stratifiedcube(X,strata,pik))
 sum(s)
 t(X/pik)\%*\%s
 t(X/pik)\%*\%pik
 
 t(Xcat)\%*\%s
 t(Xcat)\%*\%pik
 

#-------- CASE 2 pik unequal and integer in each strata  

 pik <- rep(inclusionprobabilities(runif(N/n),1),n)
 X <- as.matrix(cbind(pik,matrix(c(x1,x2),ncol = 2)))
 system.time(s <- stratifiedcube(X,strata,pik))
 sum(s)
 t(X/pik)\%*\%s
 t(X/pik)\%*\%pik
 
 t(Xcat)\%*\%s
 t(Xcat)\%*\%pik  

#-------- CASE 3 pik equal and NOT integer in each strata 
 
 pik <- rep(0.25,N)
 X <- as.matrix(cbind(pik,matrix(c(x1,x2),ncol = 2)))
 system.time(s <- stratifiedcube(X,strata,pik))
 sum(s)
 t(X/pik)\%*\%s
 t(X/pik)\%*\%pik
 
 t(Xcat)\%*\%s
 t(Xcat)\%*\%pik
 
#-------- CASE 4 pik unequal and NOT integer in each strata 
 
 pik <- rep(inclusionprobabilities(runif(N/n),2.5),n)
 X <- as.matrix(cbind(pik,matrix(c(x1,x2),ncol = 2)))
 system.time(s <- stratifiedcube(X,strata,pik))
 sum(s)
 t(X/pik)\%*\%s
 t(X/pik)\%*\%pik
 
 t(Xcat)\%*\%s
 t(Xcat)\%*\%pik


}
